WPG.17: settle Plot audit-code findings (port marker Set/Get options)
- port the marker-style Set/Get arms (ExecOptions.pas:615-682/978-1041,
  options 74-105 subset) + Get Daisysize; pinned end-to-end by the new
  circuit_marked golden
- channels= negative values now wrap like FPC Round->Cardinal
- unsolved-guard unit coverage extended to all guarded letters

+++ crates/dss-core/src/exec/plot.rs
@@ -241,7 +241,11 @@ impl Dss {
                         let n = n.min(51);
                         p.channels = buf[..n]
                             .iter()
-                            .map(|v| v.round_ties_even() as u32)
+                            // FPC `Round` (banker's) into an `array of Cardinal`:
+                            // a negative wraps modulo 2^32 (range checks off
+                            // upstream), so go through i64 — `as u32` alone
+                            // would saturate to 0 (audit settlement).
+                            .map(|v| (v.round_ties_even() as i64) as u32)
                             .collect();
                         p.bases = vec![1.0; n];
                     }
